examples/openstack/requirements.txt
python-openstackclient==5.8.0
PyYAML==6.0
